pyproject.toml
[project]
name = "myelin"
version = "0.1.0"
description = "A comprehensive Python toolkit for interacting with key components of the US healthcare reimbursement system."
readme = "README.md"
license-files = [ "LICENSE" ]
authors = [
  {name = "Jacob Wilson", email = "jjw07006@gmail.com"},
  {name = "Josh Lankford", email = "lankford.josh@gmail.com"},
]
keywords = [ "cms", "drg", "msdrg", "ioce", "mce", "ipps", "opps", "ipf", "ipsf", "ltch", "opsf", "hospice" ]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "jpype1>=1.6.0",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
    "pluggy>=1.5.0",
    "sqlalchemy>=2.0.43",
    "pytest>=8.4.2",
]

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.8",
    "pre-commit>=4.3.0",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["myelin*"]

[project.urls]
Homepage = "https://github.com/LibrePPS/myelin"
Repository = "https://github.com/LibrePPS/myelin.git"
Issues = "https://github.com/LibrePPS/myelin/issues"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
